Fix compiler warnings.

+++ ably-ios/ARTLog.m
@@ -85,7 +85,7 @@ static const char *logLevelName(ARTLogLevel level) {
 - (void)debug:(const char *)fileName line:(NSUInteger)line message:(NSString *)message, ... {
     va_list args;
     va_start(args, message);
-    [self log:[[NSString alloc] initWithFormat:[NSString stringWithFormat:@"(%@:%d) %@", [[NSString stringWithUTF8String:fileName] lastPathComponent], line, message] arguments:args] level:ARTLogLevelDebug];
+    [self log:[[NSString alloc] initWithFormat:[NSString stringWithFormat:@"(%@:%lu) %@", [[NSString stringWithUTF8String:fileName] lastPathComponent], (unsigned long)line, message] arguments:args] level:ARTLogLevelDebug];
     va_end(args);
 }
