+++ ably-ios/ARTPresenceMessage.m
@@ -29,7 +29,7 @@
     NSMutableString *description = [[super description] mutableCopy];
     [description deleteCharactersInRange:NSMakeRange(description.length - (description.length>2 ? 2:0), 2)];
     [description appendFormat:@",\n"];
-    [description appendFormat:@" action: %d\n", self.action];
+    [description appendFormat:@" action: %lu\n", (unsigned long)self.action];
     [description appendFormat:@"}"];
     return description;
 }
